Add WebGPURenderPipeline.
